ref: Switch themes again
This bases the theme off of the anpu theme, which is structurally closer
to what I am looking for, only with some small stylistic elements that
have needed to be changed. There have been a number of changes, so I am
giving it the name mals-theme distinguishing it from the original.